refactoring PostService
