+++ src/main/java/faang/school/postservice/moderation/banners/CommenterBanner.java
@@ -1,6 +1,6 @@
 package faang.school.postservice.moderation.banners;
 
-import faang.school.postservice.service.PostService;
+import faang.school.postservice.service.CommentBanService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -9,9 +9,9 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 public class CommenterBanner {
-    private final PostService postService;
+    private final CommentBanService commentBanService;
     @Scheduled(cron = "${comment.ban.scheduler.cron}")
     public void banCommenters() {
-        postService.findCommentersAndPublishBanEvent();
+        commentBanService.findCommentersAndPublishBanEvent();
     }
 }
